requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
boto3==1.16.25
botocore==1.19.25
dj-database-url==0.5.0
Django==3.1.3
django-active-link==0.1.6
django-autocomplete-light==3.8.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-storages==1.10.1
feedparser==6.0.2
gunicorn==20.0.4
jmespath==0.10.0
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.4
s3transfer==0.3.3
sgmllib3k==1.0.0
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.2
whitenoise==5.2.0
